speech_to_text/requirements.txt
whisperx
openai-whisper
numpy
huggingface_hub
requests
urllib3
google-genai
python-dotenv